Created Homepage compoenent and refactored the code
