+++ src/component/convert.js
@@ -8,6 +8,7 @@ export default function Convert(props) {
     onChangeAmount,
     amount
   } = props;
+  
   return (
     <div>
       <input type="number" className="input" value={amount} onChange={onChangeAmount}/>
